refactor : Auth Main 위치 수정
