+++ Android/app/src/main/java/com/booktree/ui/MainActivity.java
@@ -1,8 +1,9 @@
-package com.booktree;
+package com.booktree.ui;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.widget.Toast;
+import com.booktree.R;
 import com.booktree.ui.book.BarcodeScanActivity;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import androidx.appcompat.app.AppCompatActivity;
