+++ Android/app/src/main/res/layout/activity_auth.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AuthActivity">
+    tools:context=".ui.AuthActivity">
 
     <com.google.android.gms.common.SignInButton
         android:id="@+id/signInButton"
